refactor: pr-checker 拆分检查功能模块
